requirements.txt
django==1.11.3
pillow
mysqlclient
djangorestframework-jwt
djangorestframework
markdown
django-filter
django-crispy-forms
django-reversion
django-formtools
future
httplib2
six
requests
coreapi
django-cors-headers
xlwt
xlsxwriter
drf-extensions
django-redis
pycryptodome
raven